pyproject.toml
[project]
name = "catbreeddetector"
version = "0.1.0"
description = "ML project with aim to detect cat's breed by its image."
authors = [
    {name = "savanna"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kagglehub (>=0.3.12,<0.4.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "transformers (>=4.52.4,<5.0.0)"
]

[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"
wandb = "^0.19.1"
rich = "^14.0.0"
fire = "^0.7.0"
pre-commit = "^4.2.0"
mlflow = "^2.20.3"
dvc = "^3.59.1"
hydra-core = "^1.3.2"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pytest-regressions = "^2.7.0"

[tool.isort]
src_paths = ["cat_breed_detector"]
profile = "black"
line_length = 79
lines_after_imports = 2

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.7.0"
torch = "^2.5.0"
torchvision = "^0.22.0"
pytorch-lightning = "^2.5.0"
pandas="^2.2.3"
requests="^2.32.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
